refactor: Refactored components fron class to functional. Refactored styles from simple to modular
