--- src/components/AddForm/post-add-form.css
@@ -1,10 +1,0 @@
-.bottom-panel {
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-
-.bottom-panel .new-post-label {
-  width: auto;
-  flex-grow: 1;
-  margin-right: 3px;
-}
