--- src/components/List/index.js
@@ -1,2 +1,0 @@
-import PostList from './post-list';
-export default PostList;
