--- src/components/List/post-list.css
@@ -1,7 +1,0 @@
-.app-list {
-    margin-top: 50px;
-}
-.app-list .list-group-item {
-  padding: 20px 35px 10px 35px;
-  margin-top: 10px;
-}
