--- src/components/ListItem/index.js
@@ -1,2 +1,0 @@
-import PostListItem from './post-list-item';
-export default PostListItem;
