--- src/components/ListItem/post-list-item.css
@@ -1,58 +1,0 @@
-.app-list-item {
-  font-size: 1.25rem;
-}
-.app-list-item button {
-  width: 35px;
-  height: 35px;
-  margin: 3px;
-  font-size: 17px;
-  border: none;
-  cursor: pointer;
-}
-.app-list-item .btn-star {
-  color: #aeaeae;
-}
-.app-list-item .btn-trash {
-  color: red;
-}
-.app-list-item .btn-like {
-  color: #aeaeae;
-}
-.app-list-item .fa-heart {
-  text-align: center;
-  color: red;
-  -webkit-transition: 0.3s all;
-  transition: 0.3s all;
-  opacity: 0.1;
-}
-
-.app-list-item button:focus {
-  -webkit-box-shadow: none;
-          box-shadow: none;
-  outline: none;
-}
-
-.app-list-item-label {
-  display: block;
-  line-height: 35px;
-  cursor: pointer;
-  -webkit-user-select: none;
-     -moz-user-select: none;
-      -ms-user-select: none;
-          user-select: none;
-  -webkit-transition: 0.5s all;
-  transition: 0.5s all;
-}
-
-.app-list-item.like .fa-heart {
-  opacity: 1;
-  -webkit-transform: translateX(0);
-          transform: translateX(0);
-}
-
-.app-list-item.important .app-list-item-label {
-  color: #FFD700;
-}
-.app-list-item.important .btn-star {
-  color: #FFD700;
-}
