--- src/components/ListItem/post-list-item.scss
@@ -1,53 +1,0 @@
-.app-list-item {
-    font-size: 1.25rem;
-    button {
-        width: 35px;
-        height: 35px;
-        margin: 3px;
-        font-size: 17px;
-        border: none;
-        cursor: pointer;        
-    }
-    .btn-star {
-        color: #aeaeae;
-    }
-    .btn-trash {
-        color: red;
-    }
-    .btn-like {
-        color: #aeaeae;
-    }
-    .fa-heart {    
-        text-align: center;
-        color: red;
-        transition: 0.3s all;
-        opacity: 0.1;
-    }
-}
-
-.app-list-item button:focus {
-    box-shadow: none;
-    outline: none;
-}
-
-.app-list-item-label {
-display: block;
-line-height: 35px;
-cursor: pointer;
-user-select: none;
-transition: 0.5s all;
-}
-
-.app-list-item.like .fa-heart {
-    opacity: 1;
-    transform: translateX(0);
-}
-
-.app-list-item.important {
-    .app-list-item-label {
-        color: #FFD700;
-    }
-    .btn-star {
-        color: #FFD700;
-    }
-}
