+++ src/components/PostList/styles.module.css
@@ -1,0 +1,12 @@
+.list {
+  margin-top: 50px;
+}
+
+.listGroup {
+  position: relative;
+  display: block;padding: 20px 35px 10px 35px;
+  margin-top: 10px;
+  margin-bottom: -1px;
+  background-color: #fff;
+  border: 1px solid rgba(0, 0, 0, .125);
+}
