+++ src/components/PostListItem/styles.module.css
@@ -1,0 +1,22 @@
+.container {
+  display: flex;
+  justify-content: space-between;
+}
+
+.container img {
+  width: 350px;
+  box-shadow: 5px 5px 5px rgba(0, 0, 0, .25);
+  border-radius: 20px;
+}
+
+.label {
+  display: block;
+  line-height: 35px;
+  cursor: pointer;
+  user-select: none;
+  transition: all .5s;
+}
+
+.buttonsContainer {
+  
+}
