--- src/components/SearchPanel/search-panel.css
@@ -1,1 +1,0 @@
-.search-panel{margin:1rem 0}.search-panel .search-input{width:auto;-webkit-box-flex:1;-ms-flex-positive:1;flex-grow:1;margin-right:3px}
