--- src/components/SearchPanel/search-panel.js
@@ -1,29 +1,0 @@
-import React, {Component} from 'react';
-
-import './search-panel.css'
-
-export default class SearchPanel extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            term: ''
-        }
-        this.onUpdateSearch = this.onUpdateSearch.bind(this);
-    }
-    onUpdateSearch(e){
-        const term = e.target.value;
-        this.setState({term});
-        this.props.onUpdateSearch(term);
-    }
-    render() {
-        return (
-            <input
-                className = "form-control search-input"
-                type = "text"
-                placeholder="Пошук па запісах"
-                value = {this.state.term}
-                onChange = {this.onUpdateSearch}
-            />
-        )
-    }
-}
