--- src/components/SearchPanel/search-panel.scss
@@ -1,8 +1,0 @@
-.search-panel {
-    margin: 1rem 0;
-    .search-input {
-        width: auto;
-        flex-grow: 1;
-        margin-right: 3px;
-    }
-}
