+++ src/components/SearchPanel/styles.module.css
@@ -1,0 +1,5 @@
+.input {
+    display: flex;
+    width: 100%;
+    margin-right: 3px;
+}
