--- src/components/app/app.css
@@ -1,10 +1,0 @@
-.app {
-  margin: 0 auto;
-  max-width: 800px;
-}
-
-.pic {
-  width: 70%;
-  box-shadow: 5px 5px 5px rgba(0, 0, 0, 0.25);
-  border-radius: 20px;
-}
