+++ src/components/app/style.module.css
@@ -1,0 +1,8 @@
+.app {
+  margin: 0 auto;
+  max-width: 800px;
+}
+
+.searchPanel {
+  display: flex;
+}
