+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/app/';
+import App from './components/App';
 
 ReactDOM.render(<App/>, document.getElementById('root'));
